Fix UI. Remove survey activity. Remove Comments. Test Function. Errors are constraint layout.
