+++ app/src/main/res/values/strings.xml
@@ -27,7 +27,7 @@
     <string name="register_button">Register</string>
     <string name="email_hint">email</string>
     <string name="password_hint">password</string>
-    <string name="survey_prompt">You don\'t have any channels! Lets create one.\n\n
+    <string name="survey_prompt">You don\'t have any channels! Lets create one.\n
                                  Select the topics you like.\n\n</string>
 
     <string name="category_1">TV &amp; Film</string>
@@ -44,7 +44,7 @@
     <string name="google_login_type">3</string>
     <string name="google_request_code">5</string>
 
-    <string name="create">Create</string>
+    <string name="create">Create Channel</string>
     <string name="shuffle">Shuffle</string>
     <string name="no_auth">Username or password is incorrect</string>
